+++ src/main/java/com/lolduo/duo/object/entity/initialInfo/SpellEntity.java
@@ -1,0 +1,28 @@
+package com.lolduo.duo.object.entity.initialInfo;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.Id;
+import javax.persistence.Table;
+
+@Entity
+@NoArgsConstructor
+@Getter
+@Table(name = "spell")
+public class SpellEntity {
+    @Id
+    private Long id;
+    @Column
+    private String name;
+    @Column
+    private String imgUrl;
+
+    public SpellEntity(Long id, String name, String imgUrl) {
+        this.id = id;
+        this.name = name;
+        this.imgUrl = imgUrl;
+    }
+}
