+++ src/main/java/com/lolduo/duo/object/response/championDetail/ChampionDetail.java
@@ -4,4 +4,5 @@ import java.util.List;
 
 public class ChampionDetail {
     List<ResponsePerk> perkInfo;
+    List<ResponseSpell> spellInfo;
 }
