+++ src/main/java/com/lolduo/duo/object/response/championDetail/ResponseSpell.java
@@ -1,6 +1,6 @@
 package com.lolduo.duo.object.response.championDetail;
 
-import com.lolduo.duo.object.response.championDetail.sub.PerkUrl;
+import com.lolduo.duo.object.response.championDetail.sub.SpellUrl;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -8,8 +8,9 @@ import lombok.NoArgsConstructor;
 import java.util.List;
 @Getter
 @NoArgsConstructor
-public class ResponsePerk {
-    private List<PerkUrl> perkList;
+public class ResponseSpell {
+    private List<SpellUrl> spellList;
+
     @ApiModelProperty(example = "1,000게임")
     private String allCount;
     @ApiModelProperty(example = "22.22%")
@@ -17,8 +18,8 @@ public class ResponsePerk {
 
 
 
-    public ResponsePerk(List<PerkUrl> perkList,  String allCount,String winRate) {
-        this.perkList = perkList;
+    public ResponseSpell(List<SpellUrl> spellList, String allCount, String winRate) {
+        this.spellList = spellList;
         this.winRate = winRate;
         this.allCount = allCount;
     }
