+++ src/main/java/com/lolduo/duo/repository/initialInfo/SpellRepository.java
@@ -1,0 +1,7 @@
+package com.lolduo.duo.repository.initialInfo;
+
+import com.lolduo.duo.object.entity.initialInfo.SpellEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface SpellRepository extends JpaRepository<SpellEntity,Long> {
+}
